Fix: ProductOption Soft Delete 로 수정 RelatedTo: #43

+++ src/main/java/com/sparta/gaeppa/product/service/ProductOptionService.java
@@ -2,6 +2,7 @@ package com.sparta.gaeppa.product.service;
 
 import com.sparta.gaeppa.global.exception.ExceptionStatus;
 import com.sparta.gaeppa.global.exception.ServiceException;
+import com.sparta.gaeppa.members.entity.MemberRole;
 import com.sparta.gaeppa.product.dto.productOption.ProductOptionListResponseDto;
 import com.sparta.gaeppa.product.dto.productOption.ProductOptionRequestDto;
 import com.sparta.gaeppa.product.dto.productOption.ProductOptionResponseDto;
@@ -9,6 +10,7 @@ import com.sparta.gaeppa.product.entity.ProductOption;
 import com.sparta.gaeppa.product.entity.ProductOptionCategory;
 import com.sparta.gaeppa.product.repository.ProductOptionCategoryRepository;
 import com.sparta.gaeppa.product.repository.ProductOptionRepository;
+import com.sparta.gaeppa.security.jwts.entity.CustomUserDetails;
 import jakarta.transaction.Transactional;
 import java.util.List;
 import java.util.UUID;
@@ -61,10 +63,15 @@ public class ProductOptionService {
     }
 
     @Transactional
-    public void deleteProductOption(UUID optionId) {
+    public void deleteProductOption(UUID optionId, CustomUserDetails userDetails) {
 
         ProductOption productOption = productOptionRepository.findById(optionId)
                 .orElseThrow(() -> new ServiceException(ExceptionStatus.PRODUCT_OPTION_NOT_FOUND));
+
+        if (userDetails.getMemberRole() != MemberRole.MASTER && userDetails.getMemberRole() != MemberRole.MANAGER
+                && !productOption.getCreatedBy().equals(userDetails.getUsername())) {
+            throw new ServiceException(ExceptionStatus.UNAUTHORIZED);
+        }
 
         productOptionRepository.delete(productOption);
     }
